+++ src/Setup.php
@@ -1,0 +1,218 @@
+<?php
+/**
+ * Initialization of theme.
+ *
+ * @since 1.0.0
+ *
+ * @package Bcgov/Theme/Block
+ */
+
+namespace Bcgov\Theme\Block;
+
+/**
+ * Setup class setups up theme.
+ */
+class Setup {
+
+    /**
+     * Constructor.
+     */
+    public function __construct() {
+        $this->hook();
+    }
+
+    /**
+     * Sets up hooks for theme.
+     *
+     * @return void
+     */
+    private function hook() {
+        $loader = new Loader();
+		$loader->add_action( 'after_setup_theme', $this, 'bcgov_block_theme' );
+		$loader->add_action( 'wp_enqueue_scripts', $this, 'bcgov_block_theme_enqueue_scripts' );
+		$loader->add_action( 'admin_enqueue_scripts', $this, 'bcgov_block_theme_enqueue_admin_scripts' );
+		$loader->add_filter( 'get_custom_logo', $this, 'bcgov_block_theme_custom_logo' );
+		$loader->add_action( 'admin_notices', $this, 'bcgov_block_theme_dependencies' );
+		$loader->add_action( 'admin_menu', $this, 'bcgov_block_theme_menu' );
+        $loader->run();
+    }
+
+	/**
+	 * Sets up theme defaults and registers support for various WordPress features.
+	 *
+	 * @since 1.0.0
+	 *
+	 * @return void
+	 */
+	public function bcgov_block_theme(): void {
+
+		load_theme_textdomain( 'bcgov-block-theme', get_template_directory() . '/languages' );
+
+		/*
+			* Enable support for Post Thumbnails on posts and pages.
+			*
+			* @link https://developer.wordpress.org/themes/functionality/featured-images-post-thumbnails/
+			*/
+		add_theme_support( 'post-thumbnails' );
+
+		// Add support for Block Styles.
+		add_theme_support( 'wp-block-styles' );
+
+		// Add support for editor styles.
+		add_theme_support( 'editor-styles' );
+
+		// Add alignment option to toolbar options.
+		add_theme_support( 'align-wide' );
+
+		// Add support for custom logos.
+		add_theme_support( 'custom-logo' );
+
+		// Enqueue editor styles.
+		add_editor_style(
+			[
+				'./assets/css/blocks.css',
+			]
+		);
+
+		// Add support for responsive embedded content.
+		add_theme_support( 'responsive-embeds' );
+	}
+
+	/**
+	 * Enqueue scripts and styles for public website.
+	 *
+	 * @since 1.0.0
+	 *
+	 * @return void
+	 */
+	public function bcgov_block_theme_enqueue_scripts(): void {
+		$public_assets  = require_once get_template_directory() . '/dist/public.asset.php';
+		$font_assets    = require_once get_template_directory() . '/dist/font.asset.php';
+		$public_version = $public_assets['version'] ?? wp_get_theme()->get( 'Version' );
+		$font_version   = $font_assets['version'] ?? wp_get_theme()->get( 'Version' );
+		wp_enqueue_script(
+			'bcgov-block-theme-public',
+			get_template_directory_uri() . '/dist/public.js',
+			$public_assets['dependencies'] ?? [],
+			$public_version,
+			true
+		);
+		wp_enqueue_style(
+			'bcgov-block-theme-public',
+			get_template_directory_uri() . '/dist/public.css',
+			[],
+			$public_version
+		);
+		wp_enqueue_style(
+			'bcgov-block-theme-font',
+			get_template_directory_uri() . '/dist/font.css',
+			[],
+			$font_version
+		);
+	}
+
+	/**
+	 * Enqueue scripts and styles for admin.
+	 *
+	 * @since 1.0.0
+	 *
+	 * @return void
+	 */
+	public function bcgov_block_theme_enqueue_admin_scripts(): void {
+		$admin_assets  = include_once get_template_directory() . '/dist/admin.asset.php';
+		$admin_version = $admin_assets['version'] ?? wp_get_theme()->get( 'Version' );
+
+		wp_enqueue_script(
+			'bcgov-block-theme-admin',
+			get_template_directory_uri() . '/dist/admin.js',
+			$admin_assets['dependencies'] ?? [],
+			$admin_version,
+			true
+		);
+		wp_enqueue_style(
+			'bcgov-block-theme-admin',
+			get_template_directory_uri() . '/dist/admin.css',
+			[],
+			$admin_version
+		);
+	}
+
+	/**
+	 * Filter the results of get_custom_logo() if no custom logo has been chosen by injecting a default image.
+	 * <-- wp:site-logo --> uses this result internally, and so can be easily used with or without choosing a custom logo.
+	 *
+	 * @param string $html Markup from original get_custom_logo call.
+	 * @since 1.0.0
+	 *
+	 * @return string
+	 */
+	public function bcgov_block_theme_custom_logo( $html ) {
+		// If the site owner has selected a site logo using the theme customizer, use it.
+		if ( has_custom_logo() ) {
+			return $html;
+		}
+		// Otherwise, inject a default logo markup for use in wp-site-logo blocks.
+		$image = '<img src="' . get_stylesheet_directory_uri() . '/dist/images/bc_gov_logo_transparent.ced9da43.png">';
+		$html  = sprintf(
+			'<a href="%1$s" class="custom-logo-link" rel="home"%2$s>%3$s</a>',
+			esc_url( home_url( '/' ) ),
+			is_front_page() && ! is_paged() ? ' aria-current="page"' : '',
+			$image
+		);
+		return $html;
+	}
+
+	/**
+	 * Sets admin notice for missing plugin dependencies.
+	 *
+	 * @since 1.0.0
+	 *
+	 * @return void
+	 */
+	public function bcgov_block_theme_dependencies() {
+		// Checks for AIOSEO breadcrumb specific function.
+		if ( ! function_exists( 'aioseo_breadcrumbs' ) ) {
+			$className = 'error';
+			$message   = __( 'Warning: The BCGov Block Theme needs the All in One SEO plugin activated to page specific enable breadcrumb navigation.', 'bcgov_block_theme' );
+
+			printf( '<div class="%1$s"><p>%2$s</p></div>', esc_attr( $className ), esc_html( $message ) );
+		}
+	}
+
+	/**
+	 * BCGov Blocks Theme: Add theme page to the admin mmenu.
+	 *
+	 * @since 1.0.0
+	 *
+	 * @return void
+	 */
+	public function bcgov_block_theme_menu() {
+		add_theme_page(
+			esc_html__( 'BCGov Blocks Theme', 'bcgov-block-theme' ), // Page title.
+			esc_html__( 'BCGov Blocks Theme', 'bcgov-block-theme' ), // Menu title.
+			'edit_theme_options',                                     // Capability.
+			'bcgov-block-theme-info',                                // Menu slug.
+			array( &$this, 'bcgov_block_theme_page_display' )        // Callback.
+		);
+	}
+
+	/**
+	 * Display BCGov Block Theme features and instructions page.
+	 *
+	 * @since 1.0.0
+	 *
+	 * @return void
+	 */
+	public function bcgov_block_theme_page_display() {
+
+		$theme = wp_get_theme();
+
+		if ( is_child_theme() ) {
+			$theme = wp_get_theme()->parent();
+		}
+
+		require_once get_template_directory() . '/inc/core/templates/theme-info-page.php';
+
+	}
+
+}