+++ CHANGELOG.md
@@ -8,4 +8,4 @@
 – Determine SCSS/style structure for theme elements ([DESCW-465](https://apps.itsm.gov.bc.ca/jira/browse/DESCW-465))
 - Essential template files ([DESCW-467](https://apps.itsm.gov.bc.ca/jira/browse/DESCW-467))
 - Initial set up of Block Pattern integration ([DESCW-463]https://apps.itsm.gov.bc.ca/jira/browse/DESCW-463)
-- Theme-based custom block refinement – core/buttons variations and filter structure ([DESCW-497]https://apps.itsm.gov.bc.ca/jira/browse/DESCW-497)
+- Theme-based custom block refinement – core/buttons variations and filter structure – Modified functions.php to work with added Setup and action/filter Loader – Cleaned up Package/Namespacing – added Basic Cards/Card Containers ([DESCW-497]https://apps.itsm.gov.bc.ca/jira/browse/DESCW-497)